feat: Update README with more explicit Coral Dev Board flash instructions
This commit 📝 greatly expands the instructions in the README for flashing the Coral Dev Board before running the project.

Now, the README 📖 not only instructs users to install MDT, but also provides the steps to carry out the flashing process with the included bash script (`bash -x ./flash.sh`).

Furthermore, the expected console output 📄 after flashing the board and powering it is included. This addition should give users a clear idea what to expect and a way to verify that the process has executed properly.

The newly-added instructions aim at improving the user experience by making the setup process more transparent and less prone to error 👍.